requirements.txt
Flask
flask-cors
pandas
numpy
pinecone
fuzzywuzzy
python-levenshtein
python-dotenv
gunicorn
scikit-learn
requests
SpeechRecognition  
google-generativeai
pydub
sarvamai
supabase
google-genai
pytest
pytest-cov
